+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## v2.0.0
+
+### Nov. 17, 2021
+
+#### Added
+
+- Ability to create a new analytical space
+- Ability to delete an analytical space
+
+### Modifications
+
+- All Entity operations require an analytical space id
+- Method names changed
+
+---
+
 ## v1.4.0
 
 ### Nov. 04, 2021
@@ -15,6 +31,8 @@
 - `Event`
 - `Smart Configuration`
 
+---
+
 ## v1.3.0
 
 ### Sep. 28, 2021
@@ -22,6 +40,8 @@
 #### Added
 
 - `queryConfiguration` Method
+
+---
 
 ## v1.2.0
 
@@ -32,6 +52,8 @@
 - `deleteEvent` Method
 - `deleteEntity` Method
 - Optional Event Parameters `id`, `startTime`, `endTime`
+
+---
 
 ## v1.1.0